Haven't implemented work history, finished the rest.

+++ FreeLancers4/Controllers/ProjectsController.cs
@@ -7,16 +7,20 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using FreeLancers4.Data;
 using FreeLancers4.Models;
+using Microsoft.AspNetCore.Identity;
+using FreeLancers4.Areas.Identity.Data;
 
 namespace FreeLancers4.Views.Projects
 {
     public class ProjectsController : Controller
     {
         private readonly FreeLancers4Context _context;
+        private readonly UserManager<FreeLancers4User> _userManager;
 
-        public ProjectsController(FreeLancers4Context context)
+        public ProjectsController(FreeLancers4Context context, UserManager<FreeLancers4User> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         // GET: Projects
@@ -82,14 +86,33 @@ namespace FreeLancers4.Views.Projects
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,ProjectTitle,PostDate,Description,Techneeded,Skills,Price,TimeFrame,DueDate,contactEmail")] Work work)
+        public async Task<IActionResult> Create([Bind("ProjectTitle,Description,Skills,Price,DueDate,ContactEmail")] Work work)
         {
+
+            if (!(await _userManager.GetUserAsync(User)).Role.Equals("client"))
+            {
+                return NotFound();
+            }
+            
+            // Minutes must be at least 10 after now
+            if((work.DueDate - DateTime.Now).TotalMinutes < 30)
+            {
+                ModelState.AddModelError("DueDate", "The time set must be at least 30 after now");
+            }
+
+            // Add other missing details during creation of the job
+            work.Owner =  await _userManager.GetUserAsync(User);
+            work.PostDate = DateTime.Now;
+            work.Skills = work.Skills.ToUpper();
+            work.WorkStatus = "Unassigned";
+
             if (ModelState.IsValid)
             {
                 _context.Add(work);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+
             return View(work);
         }
 
@@ -100,13 +123,27 @@ namespace FreeLancers4.Views.Projects
             {
                 return NotFound();
             }
+
+            FreeLancers4User user = await _userManager.GetUserAsync(User);
+
+            if (!user.Role.Equals("client"))
+            {
+                return NotFound();
+            }
+
             ViewBag.Text = projecttitle;
 
-            var work = await _context.Work.FindAsync(id);//changes the name at the url place, //cant chnage it :(
+            var work = await _context.Work.FindAsync(id);//changes the name at the url place
             if (work == null)
             {
                 return NotFound();
             }
+
+            if (!work.Owner.Equals(user))
+            {
+                return NotFound();
+            }
+
             return View(work);
         }
 
@@ -115,9 +152,16 @@ namespace FreeLancers4.Views.Projects
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,ProjectTitle,PostDate,Description,Techneeded,Skills,Price,TimeFrame,DueDate, contactEmail")] Work work)
+        public async Task<IActionResult> Edit(int id, [Bind("ProjectTitle,Description,Skills,Price,DueDate,ContactEmail")] Work work)
         {
             if (id != work.ID)
+            {
+                return NotFound();
+            }
+
+            FreeLancers4User user = await _userManager.GetUserAsync(User);
+
+            if (!work.Owner.Equals(user))
             {
                 return NotFound();
             }
@@ -142,6 +186,7 @@ namespace FreeLancers4.Views.Projects
                 }
                 return RedirectToAction(nameof(Index));
             }
+
             return View(work);
         }
 
@@ -155,7 +200,15 @@ namespace FreeLancers4.Views.Projects
 
             var work = await _context.Work
                 .FirstOrDefaultAsync(m => m.ID == id);
+
             if (work == null)
+            {
+                return NotFound();
+            }
+
+            FreeLancers4User user = await _userManager.GetUserAsync(User);
+
+            if (!work.Owner.Equals(user))
             {
                 return NotFound();
             }
@@ -169,8 +222,68 @@ namespace FreeLancers4.Views.Projects
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var work = await _context.Work.FindAsync(id);
+
+            FreeLancers4User user = await _userManager.GetUserAsync(User);
+
+            if (!work.Owner.Equals(user))
+            {
+                return NotFound();
+            }
+
             _context.Work.Remove(work);
             await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
+        }
+
+        [HttpPost, ActionName("Accept")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Accept(int id)
+        {
+            var work = await _context.Work.FindAsync(id);
+
+            if (work == null)
+            {
+                return NotFound();
+            }
+
+            FreeLancers4User user = await _userManager.GetUserAsync(User);
+
+            if(work.AssignedTo != null)
+            {
+                ModelState.AddModelError("Assigned", "The job has already been assigned to someone else.");
+            } 
+            else
+            {
+                work.Assigned = user;
+                work.WorkStatus = "Assigned";
+            }
+
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    _context.History.Add(new History {
+                        Project = work,
+                    });
+
+                    _context.Update(work);
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!WorkExists(work.ID))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+
+                return RedirectToAction(nameof(Index));
+            }
+
             return RedirectToAction(nameof(Index));
         }
